erradicated JDs fetch value code and turned to geMotion6. Finishing debugging.

+++ catkin/src/tsn_imu_base/tsn_mpu6050/include/ros_mpu6050.h
@@ -67,7 +67,7 @@ private:
 
 	sensor_msgs::Imu data_out;
 
-	double ax, ay, az, gx, gy, gz;
+	int16_t ax, ay, az, gx, gy, gz;
 };
 
 #endif
